// frontend/src/app/model/weight/tanita-measurement.model.ts
export interface TanitaMeasurementModel {

  date: Date;

  weight: number;

  bmi: number;

  bodyFatRatio: number;

  bodyFatMass: number;

  visceralFat: number;

  muscleMass: number;

  muscleMassRatio: number;

  muscleQuality: number;

  boneMass: number;

  bmr: number;

  metabolicAge: number;

  bodyWatterRatio: number;

  bodyWatterMass: number;

  physiqueRating: number;
}